Portfolio/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.2
uvicorn==0.32.1
